# cm.domeni.kapita/openapi/main.yaml
---
openapi: "3.0.3"
info:
  title: "kapita"
  description: "REST API endpoints for kapita"
  version: "v-000"
servers:
- url: "https"
- url: "http"
tags:
- name: "Demo"
  description: "Demo operations"
- name: "Debt"
  description: "Debt operations"
- name: "Transaction"
  description: "Transaction operations"
paths:
  /demo:
    get:
      tags:
      - "Demo"
      summary: "Fetch all Demos"
      description: "Fetch all Demos"
      operationId: "fetchAllDemo"
      security:
      - oAuth:
        - "demo:read:all"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DemoDTO"
    post:
      tags:
      - "Demo"
      summary: "Create a new Demo"
      description: "Create a new Demo"
      operationId: "createDemo"
      security:
      - oAuth:
        - "demo:create"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDemoDTO"
      responses:
        "201":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponseDTO"
  /demo/{demoId}:
    get:
      tags:
      - "Demo"
      summary: "Fetch demo by id"
      description: "Fetch demo by id"
      operationId: "fetchDemoById"
      security:
      - bearerAuth: []
      parameters:
      - name: "demoId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
        description: "represents the id of demo to fetch."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoDTO"
        "404":
          description: "KAPITA-404-001"
  /transaction:
    get:
      tags:
      - "Transaction"
      summary: "Fetch current user transactions"
      description: "Fetch current user transactions with optional type filtering and pagination"
      operationId: "fetchTransactions"
      security:
      - bearerAuth: []
      parameters:
      - name: "type"
        in: "query"
        required: false
        schema:
          $ref: "#/components/schemas/TransactionTypeDTO"
        description: "Transaction type to filter. When omitted, all transaction types are returned."
      - name: "pageNumber"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          minimum: 0
          default: 0
        description: "Zero-based page number. Defaults to 0."
      - name: "pageSize"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          minimum: 1
          default: 10
        description: "Number of items per page. Defaults to 10."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPageDTO"
    post:
      tags:
      - "Transaction"
      summary: "Create a new transaction"
      description: "Create a new transaction"
      operationId: "createTransaction"
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionDTO"
      responses:
        "201":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponseDTO"
  /debt:
    get:
      tags:
      - "Debt"
      summary: "Fetch current user debts"
      description: "Fetch current user debts with optional type filtering and pagination"
      operationId: "fetchDebtsByType"
      security:
      - bearerAuth: []
      parameters:
      - name: "type"
        in: "query"
        required: false
        schema:
          $ref: "#/components/schemas/DebtTypeDTO"
        description: "Debt type to filter. When omitted, all debt types are returned."
      - name: "pageNumber"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          minimum: 0
          default: 0
        description: "Zero-based page number. Defaults to 0."
      - name: "pageSize"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          minimum: 1
          default: 10
        description: "Number of items per page. Defaults to 10."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPageDTO"
    post:
      tags:
      - "Debt"
      summary: "Create a new debt"
      description: "Create a new debt for the current user"
      operationId: "createDebt"
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtDTO"
      responses:
        "201":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponseDTO"
  /debt/{debtId}/paid:
    put:
      tags:
      - "Debt"
      summary: "Mark a debt as paid"
      description: "Mark the current user debt as paid and create the related transaction"
      operationId: "markDebtAsPaid"
      security:
      - bearerAuth: []
      parameters:
      - name: "debtId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
        description: "Debt identifier."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtDTO"
        "404":
          description: "KAPITA-404-002"
  /transaction/balance:
    get:
      tags:
      - "Transaction"
      summary: "Fetch current user balance over a period"
      description: "Compute the current user balance over a period using incoming transactions minus expenses"
      operationId: "fetchTransactionBalance"
      security:
      - bearerAuth: []
      parameters:
      - name: "startDate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "date"
        description: "Inclusive period start date."
      - name: "endDate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "date"
        description: "Inclusive period end date."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyDTO"
  /transaction/amount:
    get:
      tags:
      - "Transaction"
      summary: "Fetch current user transaction amount by type over a period"
      description: "Compute the sum of transaction amounts for the current user, filtered by type and period"
      operationId: "fetchTransactionAmountByType"
      security:
      - bearerAuth: []
      parameters:
      - name: "startDate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "date"
        description: "Inclusive period start date."
      - name: "endDate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "date"
        description: "Inclusive period end date."
      - name: "type"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/TransactionTypeDTO"
        description: "Transaction type to aggregate."
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyDTO"
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    oAuth:
      type: "oauth2"
      description: "This API uses OAuth 2 with the Authorization Code grant flow."
      flows:
        authorizationCode:
          authorizationUrl: ""
          tokenUrl: ""
          scopes:
            demo:read:all: "Needed to fetch all Demo objects"
            demo:create: "Needed when to create a new Demo object"
            transaction:create: "Needed when to create a new transaction"
            transaction:read:balance: "Needed when to fetch the current user balance over a period"
            transaction:read:amount: "Needed when to fetch the current user transaction amount over a period for one type"
            debt:create: "Needed when to create a new debt"
            debt:read: "Needed when to fetch current user debts by type"
  schemas:
    DemoDTO:
      type: "object"
      required:
      - "id"
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
    CreateDemoDTO:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
    CreationResponseDTO:
      type: "object"
      required:
      - "newId"
      properties:
        newId:
          type: "string"
          format: "uuid"
    DebtTypeDTO:
      type: "string"
      enum:
      - "RECEIVABLE"
      - "PAYABLE"
    DebtStatusDTO:
      type: "string"
      enum:
      - "UNPAID"
      - "PAID"
    CreateDebtDTO:
      type: "object"
      required:
      - "type"
      - "counterpartyName"
      - "amount"
      - "dueDate"
      properties:
        type:
          $ref: "#/components/schemas/DebtTypeDTO"
        counterpartyName:
          type: "string"
          minLength: 1
          pattern: ".*\\S.*"
        amount:
          $ref: "#/components/schemas/MoneyDTO"
        dueDate:
          type: "string"
          format: "date"
    DebtDTO:
      type: "object"
      required:
      - "id"
      - "type"
      - "counterpartyName"
      - "amount"
      - "dueDate"
      - "status"
      - "createdAt"
      properties:
        id:
          type: "string"
          format: "uuid"
        type:
          $ref: "#/components/schemas/DebtTypeDTO"
        counterpartyName:
          type: "string"
        amount:
          $ref: "#/components/schemas/MoneyDTO"
        dueDate:
          type: "string"
          format: "date"
        status:
          $ref: "#/components/schemas/DebtStatusDTO"
        createdAt:
          type: "string"
          format: "date-time"
    DebtPageDTO:
      type: "object"
      required:
      - "items"
      - "pageNumber"
      - "pageSize"
      - "totalElements"
      - "totalPages"
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/DebtDTO"
        pageNumber:
          type: "integer"
          format: "int32"
          minimum: 0
        pageSize:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        totalPages:
          type: "integer"
          format: "int32"
    TransactionTypeDTO:
      type: "string"
      enum:
      - "INCOMING"
      - "EXPENSE"
    TransactionCategoryDTO:
      type: "string"
      enum:
      - "STOCK"
      - "TRANSPORT"
      - "FOOD"
      - "PHONE_CREDIT"
      - "RENT"
      - "EMPLOYEE_SALARY"
      - "SALE"
      - "SERVICE_PROVIDED"
      - "SALARY"
      - "OTHER"
    CreateTransactionDTO:
      type: "object"
      required:
      - "type"
      - "category"
      - "amount"
      properties:
        type:
          $ref: "#/components/schemas/TransactionTypeDTO"
        category:
          $ref: "#/components/schemas/TransactionCategoryDTO"
        otherCategoryDetail:
          type: "string"
          description: "Required when category is OTHER"
        amount:
          type: "number"
          minimum: 0
          exclusiveMinimum: true
        description:
          type: "string"
    TransactionDTO:
      type: "object"
      required:
      - "id"
      - "type"
      - "category"
      - "amount"
      - "createdAt"
      properties:
        id:
          type: "string"
          format: "uuid"
        type:
          $ref: "#/components/schemas/TransactionTypeDTO"
        category:
          $ref: "#/components/schemas/TransactionCategoryDTO"
        otherCategoryDetail:
          type: "string"
        amount:
          type: "number"
        description:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
    TransactionPageDTO:
      type: "object"
      required:
      - "items"
      - "pageNumber"
      - "pageSize"
      - "totalElements"
      - "totalPages"
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/TransactionDTO"
        pageNumber:
          type: "integer"
          format: "int32"
          minimum: 0
        pageSize:
          type: "integer"
          format: "int32"
          minimum: 1
        totalElements:
          type: "integer"
          format: "int64"
        totalPages:
          type: "integer"
          format: "int32"
    MoneyDTO:
      type: "object"
      required:
        - "value"
        - "currency"
      properties:
        value:
          type: "number"
        currency:
          type: "string"
          minLength: 3
          maxLength: 3
          pattern: "^[A-Z]{3}$"